Add _shared.utils._set_order function
